replicators: Use "snapshotting" in logs for snapshotting
We've been pushing a lot to get more consistent about the language we
use around snapshotting (the process of initially copying all the schema
and data for a database over to ReadySet) and replicating (the process
of forwarding changes made to the upstream database to ReadySet). A
couple of old log spans and events still used the word "Replicating" for
a process that was actually snapshotting, specifically the snapshotting
of the data in individual tables - this updates those to use the correct
word instead.

Change-Id: I599f3927f29ddf4612a8df1e14f35dfeedf5c122
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3968
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI