replicators: Mark rels as non-replicated during replication
If something fails when creating a table or a view during replication,
mark it as non-replicated in the server. This takes advantage of the
fact that we can now parse the *names* of both CREATE TABLE and CREATE
VIEW statements without being able to parse the actual body of the
statements.

Release-Note-Core: Fixed a bug where tables and views that failed to
  replicate from the upstream database would not participate in schema
  resolution, which could cause certain queries with
  non-schema-qualified table references to read from a table in the
  wrong schema rather than being proxied to the upstream database.
Fixes: ENG-2134
Change-Id: I2a2aeb7bef367b1d3a2f4d0d62a4da175b10b177
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3972
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>