readyset-dataflow: Handle failed base table ops
If we fail to apply operations to base tables during replication, that
means we have missed some changes and have invalid state, so this change
returns errors in those cases.

If we _aren't_ replicating because we don't have an upstream db, we want
to allow failed updates and deletes to base tables in order for our
rocks db tables to act like an upstream--if you remove or update a
non-existent row in postgres, for example, it warns about it but treats
it as a no-op. There are tests that rely on this behavior.

Refs: ENG-2183
Change-Id: Ifd31a112f9028ed0f3916b92e7cf58407950fe71
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3957
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>