sql: Push non_replicated_relations down to mir_converter
When building queries, we need to know the set of non-replicated
relations so we can return ReadySetError::TableNotReplicated instead of
ReadySetError::TableNotFound. Since this happens in SqlToMirConverter,
this commit pushes the storage of the `non_replicated_relations` field
down a level from SqlIncorporator to that struct so we have access to it
there.

Refs: ENG-2135
Change-Id: I43aa9020fa53089c05e03ef7e79b07c8635a7bff
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3974
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>