treewide: Return TableNotReplicated when appropriate
Everywhere where we currently return ReadySetError::TableNotFound for a
non-found table but have access to the set of non-replicated relations,
instead return ReadySetError::TableNotReplicated if the non-found table
exists in the set of non-replicated tables. This already has the nice
effect of making queries that trigger these errors get marked as
unsupported instead of permanently pending (since tables might just not
yet have snapshotted yet, any queries that fail to migrate due to
TableNotFound are marked as pending instead of unsupported in the
migration handler in the adapter) but will also be used to change the
log when this happens from an error into a warning (next up).

Release-Note-Core: Queries that fail to run against readyset due to
  reading from a table that ReadySet is not replicating from (either
  because the table has been configured not to be replicated, or because
  replicating the table failed) are now marked as unsupported instead of
  permanently being marked as pending.
Refs: ENG-2135
Change-Id: I67cbd62ff889211bcfc78c728c2bc80c26e42cf7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3976
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>