adapter: Log at warn-level for TableNotReplicated
Update the log for errors during query migration to log at warn-level if
the error is caused by a TableNotReplicated error, rather than
error-level, since this is not a fatal condition.  There may be other
places that we log errors during migration of queries, but this seems to
be the big one that's getting hit in practice.

Release-Note-Core: Queries that fail to run against readyset due to
  reading from a table that ReadySet is not replicating from (either
  because the table has been configured not to be replicated, or because
  replicating the table failed) now log at warn-level instead of
  error-level.
Fixes: ENG-2135
Change-Id: If151ec182ec6873ecbbd781f78acf27cc3840691
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3978
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>