Inlining: Unsupported Placeholder Detection Pass
Adds a pass to detect unsupported placeholders in a query. This pass
currently identifies as unsupported

1. any placeholder not in the WHERE clause,
2. any placeholder in a subquery,
3. any placeholder that does not appear in the right side of an equality
or ordering comparison with a column on the left side -or-,
4. any placeholder that does not appear in a between clause that
compares a column with two literals. These literals may be a mix of
non-placeholder and placeholder literals.

This pass does not capture all conditions that may make a placeholder
unsupported, but will not mark a supported placeholder as unsupported.
It does not check that the WHERE clause is in conjunctive normal form,
or whether we have mixed equality and ordering comparisons.

This pass will be run after all rewrite passes in the server, as some
rewrite passes will allow us to support a query.

Change-Id: Iec8dabcfbfebf5bff9736da708018a1eee4f5710
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4125
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>