controller: Skip non-base nodes when collecting replication offsets
A recent change caused us to sometimes end up with replication offset
entries for non-base nodes, which would sometimes (always?) be set to
None, and in some circumstances could trigger a panic due to replication
offsets missing after snapshot. For reference:

72f191c17 (dataflow: Fix replication offsets in memory state, 2022-12-01)

This fix simply adds an additional check to the controller to only
include replication offsets for base tables in the response to the
/replication_offsets request.

This commit also includes a replicator test that fails without this fix.

Release-Note-Core: Fixed a crash bug that could be triggered when using
  Postgres as the upstream database, which would show as a panic with
  reason "Maximum offset must be present after snapshot".
Fixes: ENG-2204
Change-Id: I70dbe08f3450169041ce25472e877269331e14ce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3987
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>