dataflow-expression: Fix caller tracking for `[try_]eval_expr`
Change Idc5909708f81a428d6fa1c25a547f4dc28f38cc2 made it such that we
get better error messages on failure. However, because of the panic
happening in a closure, `#[track_caller]` did not get propagated.

This commit fixes the error reporting to track the function caller by
having the panic be invoked in a `match` of the `Result` values.

Change-Id: I850aad70532b7d3d2ffd23e1920317e0c71786df
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3993
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI