Inlining: Detect and return UnsupportedPlaceholders error
Prior to attempting a migration, we now run the
detect_unsupported_placeholders on the SelectStatement. If the pass
fails, we attempt to borrow a cache for the query. If we cannot borrow a
cache, we return the UnsupportedPlaceholders error.

We no longer will attempt to reuse a cache if the unsupported
placeholder detection pass succeeds. Note that this will cause us to
miss queries that could be reused but are not caught by the detection
pass.

Change-Id: I03d5625ed8a1d1fccb34852b6e2361636843a519
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4126
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>