mir: Fixed bug in query removal logic.
The query removal logic was pretty flawed, and it worked in an
asymmetric way to query addition: for adding queries, we rely on MIR to
add new nodes to Dataflow through a `Migration`; for removing queries,
we just get the leaf node (for queries/views) or root node (for tables),
 and then we manually remove that node through a `Migration` outside of
MIR scope.
Now, the MIR code handles the removal (registering removed nodes in a
`Migration`) as well as notifying which relations were removed, which is
 now used by the `Recipe` to clean-up its registry.

The algorithm to traverse and remove MIR nodes also needed a change.
Roughly put, we were always traversing the MIR graph downwards from a
given node towards the leaves, removing the nodes that we encountered in
 the middle and assuming that was enough. The problem with that approach
 is that it left MIR in an inconsistent state, since JOINs and UNIONs
 queries were only pruned, but not completely removed.
The new algorithm now looks for the leaves first (and keeps track of all
 the queries that own any node in the way), and then traverses the MIR
graph backwards, from leaves to roots, removing the affected queries
from the ownership of any node in the middle, and removing the nodes if
 there were no other owners left (this last part does not apply to base
table nodes).

Refs: ENG-1634
Change-Id: I17cd9de289283e440a071b6c192ae7b942054d6b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3905
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>