sql: Store relevant data from the stmt in the QueryGraph
Rather than converting *most* of the SelectStatement to a QueryGraph,
but then passing around both the statement and the QueryGraph
everywhere, just store all the extra information we need from the
SelectStatement on the QueryGraph itself, so that it becomes possible to
compile a query to MIR entirely from a QueryGraph. This is something
we've been wanting to do for a *while* for several reasons - it means
that we're using the AST directly in fewer places (which reduces the
cost of the AST changing, as it frequently does) and it makes the
compilation pipeline a lot clearer cognitively (now we're very clearly
just String -> AST -> QueryGraph -> MIR -> Dataflow). The reason I'm
doing it *now*, though, is that we'd like to avoid actually making nodes
for views until a cached query selects FROM those views (ENG-2136) - and
the most convenient thing to store for an "unmigrated" view is in fact
the query graph - this allows us to *just* store the query graph, and
not worry about also storing the select statement.

I took this opportunity to refactor the `to_query_graph` function
slightly, to avoid local mutation and instead construct the query graph
itself at the end - this is slightly better style since it ensures we
don't forget to initialize any fields. Other than that, the only thing
that changed there is the initialization of the new fields that were
added (including `is_correlated` and `default_row`, since both are done
syntactically anyway).

The only wart here is that we have to store both the original `fields`
of the SelectStatement *and* the vector of `OutputColumn` we were
originally storing - I tried to make this work by just doing the leaf
projected fields via the output columns, but that failed due to some
really tricky-to-follow MIR column aliasing/naming issues that I don't
want to fully track down right now. I've left a TODO in the code to look
at that next time we're touching QueryGraph, but for now this isn't
actually *worse* since we were already basing the leaf projected columns
on the `fields` of the statement; now, we're just loading the
information from a different place.

Change-Id: I4458c766dd8fe2bbb76266472710ff9cd101a44a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4007
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>