mir: Set `DfNodeAddress` to `None` for leaf and alias MIR nodes.
The leaf and alias MIR nodes do not have a Dataflow node counterpart. We
 were previously setting their parent's Dataflow node as theirs, but
that made it hard to get the removal logic right.
To understand why, consider this scenario: we have a base table and a
query (no matter what that query looks like) that depends on that table.
In MIR, we'd have an alias node immediately below the base table.
Then, if we decide to remove the query (and for that reason, grab all
the MIR nodes that need to be removed), we will remove the alias node,
but we should NOT delete its Dataflow node (since its Dataflow node is
its parent's Dataflow node, which is the base table!).

Now, if we need the DF nodes for leaf/alias nodes, we resolve them by
traversing the MIR graph from those nodes towards the roots, until we
find the last DF node assigned.

This commit also introduces a new field to `MirNodeInner::Leaf`, to keep
 track of whether the node was already lowered to DF or not.
We need this, since otherwise we will process the leaf node more than
once.

Refs: ENG-1634
Change-Id: I0130453388c9fcf8ba60c6bed428bf87903b3454
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4014
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>