mir: Better differentiate between MIR and DF node indexes
Now that we're using a petgraph Graph for MIR nodes, it's much less
clear when passing around a `NodeIndex` whether that is a node index for
a MIR node or a node index for a dataflow node. This commit clears that
up at the type-level, by defining an alternate internal `Ix` for MIR
which wraps the `u32` we were already using for our Ix, but is a
distinct type so we are guaranteed not to mix up dataflow and MIR node
indexes.

Change-Id: Id29f900071f63da89538d7d863726cb60acb6e72
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4010
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>