replicators: Move version query before ddl setup
If we are connecting to postgres 13, the ddl setup commands we run cause
our connection to be considered a "replication connection", which if the
upstream is configured in a certain way (in a way that Aurora Postgres
is configured), would cause the query for the server version to fail.

Moving the query to before the ddl setup allows the query to succeed.

Release-Note-Core: Fixed a bug that would cause replication to fail on
  Aurora Postgres version 13 databases.
Refs: ENG-2208
Change-Id: Ie3f506e593c9bbd11802852472b0d29cbc5b0671
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4015
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>