replicators: Use caches, not views, for tests
It's soon going to become the case that we don't create leaf reader
nodes for views, since under normal (SQL) operation it's impossible to
actually *read* from those reader nodes (the user has to create a cache
that does a `SELECT *` from them, at which point we'll make a new reader
node for that cache). This is mostly fine, except that the test suite
for the replicators crate currently depends on the old behavior of
creating leaf nodes implicitly for views. In preparation for that
change, this commit refactors those tests such that they instead
explicitly create caches that select from those views when checking
results.

Refs: ENG-2136
Change-Id: I394aeb1dc4bb7f347b7abeebdab040c69e6b6476
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4016
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>