server/integration: Remove low-value, low-signal tests
There are a bunch of tests in server/integration that don't test things
that are good fits for server integration tests, specifically in this
case especially deep combinations of various operators, which are much
better suited to testing via logictests (and are in fact already covered
by logictests!). To save on the time we spend every build compiling and
running these tests, let's just remove them.

Refs: ENG-2186
Change-Id: Id8a2c3293c3be6cf23f56200e97c404e75da9a79
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4030
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>