sql: Don't create nodes for views until a query references them
When new VIEWs are added to the recipe, don't actually migrate them to
nodes - instead, store them in a new `unmigrated_views` map inside of
`SqlIncorporator`. Then, when a query we're migrating for any reason
references one of these views, actually turn it into nodes and remove it
from the map.

At this stage, the primary win is efficiency - there's no point in
creating nodes for views which nobody will ever actually query. However,
this paves the road to actually fixing a bug where views which select
from *each other* that get snapshotted out of order fail to migrate -
actually fixing that, however, requires moving query rewriting back into
SqlIncorporator - which in turn requires moving ExprRegistry into
SqlIncorporator and (ideally) requires removing Recipe.

Also, because it's not *currently* possible to register a leaf relation
for a MIR query without also creating a reader node for it, this *does*
still create a (fully materialized!!) reader node for views which nobody
can ever actually perform reads against - fixing that is up next.

Fixes: ENG-2136
Change-Id: I157fa5c6cabd5aaf12e17c6351e6389dfcfbbe66
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4035
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>