replicators: Ignore table that see err in repl
If we succeed in snapshotting a table but later see an error replicating
that table, it is better to ignore that table retroactively and keep
working than get stuck in a replication failure loop.

We don't expect to happen generally, but have seen
one case recently with generate columns--there could be others.

If we see this error, it means we should fix the underlying issue either
by ignoring that category of table before snapshotting, or fixing the
thing preventing us from replicating.

Release-Note-Core: Tables that encounter an error in replication will
  now be removed from the set of replicated tables automatically.
Change-Id: I7b99c8729b5e9d23433f8c470c11a968526f146e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3491
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>