rewrite: Fixed LIMIT ?, ? params getting swapped.
At the AST level, we were incorrectly normalizing the `limit` and
`offset` information.
For that reason, the AST didn't give us correct information about the
syntax that was being used, which could either be `LIMIT <offset>,
<limit>` (only available in MySQL) or the standard `LIMIT <limit> OFFSET
 <offset>`.
That was the cause of the bug swapping the parameters for `LIMIT ?, ?`,
since we wrongly assumed that the last parameter (in every case) would
always be the `offset`, which is not true for this syntax.

In order to solve this, the `limit` and `offset` fields in the AST have
been unified under a `limit_clause`, where we have an enum to
distinguish between the two possible syntaxes.
Most of the changes are just a find-and-replace current usages of those
fields for this new one.

A couple of new scenarios have been added to the `correct_offset_limit`
test in `rewrite.rs` to ensure we correctly rewrite queries using
the `LIMIT ?, ?` syntax (and that we correctly get the right `limit` and
 `offset` params).

Refs: ENG-1589
Change-Id: I2efbf386abec2c881cfc531054f4c9530c0969be
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4033
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>