launchpad: Add new panic-catching form for `eventually!` macro
This code extends the existing `eventually!` macro by adding a new form
that specifies two separate code blocks which are labeled `run_test:`
and `then_assert:`, which are used for running the test and for
validating the results separately. Panics in the `then_assert` block
will be caught, and both blocks will then be retried after a delay.

This construct allows the test to still fail immediately if an
unexpected panic occurs during the main test, but lets us use unwraps
and the `assert!` family of macros to validate results without giving up
the retry logic afforded by `eventually!`. In many cases, this lets us
write tests that are more readable and produce better error messages.

(Note also that code in the `then_assert` block must be unwind-safe,
which is typically easy to achieve for simple unwraps and asserts, but
is often not possible for the kind of code we run in tests prior to
checking results. Thus, another reason to require two separate code
blocks and only catch panics in the second block.)

One rough corner (which I've pointed out in the doc comments) is that
the macro is written to use the same syntax as closures for the
`then_assert` parameter at the end, but contrary to appearances we are
not actually creating a closure here, so we cannot pass in a closure
value as an argument. Since this is somewhat counter-intuitive, I'm
planning to add some extra code in a followup commit to provide a
helpful compile error if anyone attempts to call this macro in a way
that might indicate an attempt to pass in a closure.

Change-Id: Icb166274596446335dcae26ffc4d4c81d0a245d8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4037
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>