launchpad: Add compile error for bad usage of `eventually!`
As noted in the docs and the previous commit message, there is a gotcha
in the `run_test`/`then_assert` form of the `eventually!` macro where it
looks like you pass it a closure, but no closure is actually created.
You cannot actually pass a variable containing a closure since the macro
sneakily converts the closure syntax to an async block. This commit
helps users avoid confusion by failing compilation with an informative
error message if it seems like the macro might be being used
incorrectly.

(Note that compilation would already have failed previous to this
commit, just with a more confusing error message.)

Change-Id: Ifd8b651dd00212ce8c396ec1b668b3602191aa83
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4038
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>