Add CommandCompletePassThrough and write row
Previously we would try to invert SELECT from INSERT when getting a
CommandComplete message using logic that was faulty in the case of a
SELECT that returns no results.

In that case, we have to write back the full select tag, along with a
preceeding row description for our behavior to match that of postgres.

This commit works in tandem with a recent change to the readysettech
fork of rust-postgres that allows for the correct behavior, but
including a new way to pass through the raw tag bytes from the
CommandComplete message, along with ensuring we write the row
description.

Release-Note-Core: Fixes an issue where we would not correctly write out
the row description or correct tag for command complete messages when
the user was issuing a SELECT that returned no results.

Change-Id: I79679adf424d37c8f4caa32354d83f33893e9563
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4040
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>