sql: Don't create leaf nodes for VIEWs
Since under normal SQL operation lookups can never be performed on
VIEWs (the user has to make a cache that selects FROM that view), we'd
like to avoid creating reader nodes for them - but we still have to
register the leaf node for the view under its name (so it can be
returned from `SqlToMirConverter::get_relation`). Previously, the
behavior of creating materialized reader nodes for a migrated query was
tightly coupled to the behavior of registering the leaf node for that
query under its name - this commit decouples those, by converting the
`is_leaf` boolean into a new `LeafBehavior` enum, with a variant to
create a reader and register the query, a variant to just register the
query but not create a reader (used for VIEWs, now), and a variant to do
neither. This not only saves memory and compute in creating and
maintaining the state of an extra fully materialized reader node, it
also avoids forcing certain internal nodes within the graph to be fully
materialized!

Fixes: ENG-2238
Change-Id: I86ecb62b2665e985fd972ed32b13676707786e38
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4042
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>