replicators: Log errors that cause adapter restarts
If we encounter an error while the adapter is running, we bubble it up
and re-start the adapter loop, but don't log the reason the adapter is
restarting. This adds such a log to help with knowing why an adapter
restarted. It is a warn log instead of an error log because there are
valid reasons to need to retry (timeouts, for example), and we expect
success on retry for a large number of them.

Change-Id: Idb94e55109837715ca8126a3824c8f3b69c880a3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4046
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>