server: Remove integration tests that depend on Recipe
We're going to be getting rid of the Recipe struct in favor of just
storing the SqlIncorporator (or, the same thing with a different name
perhaps) in the DfState. Some integration tests, are tightly coupled to
the Recipe itself - this removes those in preparation, because they're
not going to be very useful afterwards.

Change-Id: I03eacbd0ced09101244de1a35163f20868c383f0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4053
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>