sql: Move ExprRegistry into SqlIncorporator
Currently, the Recipe struct is "just" a SqlIncorporator plus an
ExprRegistry. Because that ExprRegistry lives outside of
SqlIncorporator, but needs to see pre-rewrite queries, we need to
rewrite queries before they end up in SqlIncorporator - which means
among other things that lazily migrated views that we receive
out-of-order get rewritten before they get a chance to see that each
other exists (which matters in the cases that views select from other
views). This commit *starts* paving the road to rewriting inside
SqlIncorporator, by simply moving the `ExprRegistry` from a field of
Recipe to a field of SqlIncorporator. All the logic is still happening
in Recipe, though.

Refs: ENG-2246
Change-Id: Ida0cd01450f38ce20fe93f89f73488fd3276537d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4055
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>