sql: Use LeafBehavior for compound queries
Use the same LeafBehavior enum introduced in 52de139d7 (sql: Don't
create leaf nodes for VIEWs, 2022-12-19) for compound queries in
addition to regular select statements. I'm still not happy with how
compound queries are designed, but I'm not going to be able to avoid
them as part of lazy view migration.

Since this has the effect of making views containing unions also not get
reader nodes, one test that depended on that had to also be updated to
use caches instead.

Change-Id: Ic004179f44fa3a045d39932de08b8c31af76f916
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4059
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI