recipe: Fixed ALTER TYPE flaw in node removal logic.
For `ALTER TYPE`s, whenever a table is affected, we drop all queries
that depended on it.
That removal logic was flawed, since we were traversing the DF graph
from the base table node (the one affected) to the leaf nodes, removing
all nodes in the way.
That process had two problems:
1. We are not removing the MIR nodes from the affected queries, so MIR
becomes inconsistent (we will still see the queries in MIR).
2. By just traversing the graph downwards and removing all nodes, we
might be just prunning some query trees. Think of a JOIN for example,
that has a base table affected by the `ALTER TYPE` and another base
table that was unaffected, as root nodes. We'll just be prunning one
branch of that query, leaving all the nodes from the unaffected base
table all the way to the join node.

Refs: ENG-2151
Change-Id: I704074febe5e73718b4ab3b40743618da9034eab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4068
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>