sql: Rewrite queries later in the compilation pipeline
When compiling queries, wait to rewrite until we're about to compile the
query to MIR - and do that *inside* the loop where we catch
table-not-found errors for uncompiled views and handle them by compiling
the view. This has the effect that views which reference each other but
are migrated out-of-order (which can happen during snapshotting, since
we don't topologically order views) still migrate successfully when the
outermost view is eventually referenced by a cache.

There were a couple of places in query_graph where we were returning
errors that were *effectively* table-not-found errors, but
not *actually* `ReadySetError::TableNotFound` and hence weren't actually
able to handle them - it turns out that if we just remove these errors,
we'll return TableNotFound later still, so this commit does that.

This all comes with a test where we have a view referencing another view
added out-of-order via extend_recipe, which checks that a query added
which references the outermost view successfully compiles.

Release-Note-Core: ReadySet now always successfully snapshots SQL VIEWs
  which select from other VIEWs
Fixes: ENG-2132
Change-Id: I23ad69630e443ff1fbf458839d4e52fa0b8c617a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4060
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>