adapter: Post-lookup filter all keys within compound open ranges
The post-filtering for compound ranges added in 54818009e (noria-client:
Post-filter compound ranges, 2022-02-09) skipped the first key when
adding post-lookup filters for compound ranges - this is the correct
behavior for closed (inclusive) ranges, but for open (exclusive) ranges
we need to filter all keys (eg because (1, 2) > (1, 1) even though
NOT (1 > 1 && 2 > 1)). This implements that behavior in
build_view_query, covering it with a unit test and a new query in the
ranges logictest which originally failed.

Fixes: ENG-2248
Change-Id: I510809c7724eebff5baf771ec7bb1ff53e538a99
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4080
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>