clustertests: Use DatabaseConnection instead of mysql::conn
This hooks up the clustertests to use DatabaseConnection instead of
mysql::Conn, which is one step closer to being able to use
DatabaseConnection::PostgreSQL in the testing setup.

Rather than refactor the existing tests to use the dialect-agnostic API,
this test dispatches to the inner mysql conn in most cases.

Actually allowing for postgres connections to be made and used will come
in a follow-up.

Change-Id: I402f72f2f283bb4ae3e42a1bb66b11fa82f4f903
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4067
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>