dataflow-expression: Implement array_to_string function
Implement the PostgreSQL array_to_string builtin function, which joins
the (innermost) elements of a (potentially multidimensional) array into
a string with a delimiter, and an optional string to use to represent
NULL elements (if the latter isn't provided, NULL elements are skipped
entirely). This is generally useful in the interest of supporting more
builtin functions, but I'm adding it now because I want to split out
GroupConcat into an array_agg followed by a project node that calls this
function, in the interest of making all aggregates actual algebraic
groups.

Release-Note-Core: Added support for the PostgreSQL `array_to_string`
  function
Change-Id: I42a886b8c402ca997457648860b324a5c2e0349a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4091
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>