controller: Add RPC to load table statuses
Add an RPC to load the status of *all* tables, including non-replicated
ones. The "status" here currently consists of an enum indicating
whether the table is not replicated, snapshotting, or fully snapshotted.

The idea here is to surface this information to the user, via a new
`SHOW READYSET TABLES` custom SQL command.

Refs: ENG-2181
Change-Id: Ie84e2c56da1a79134430958a7abb5b70d46da4e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4098
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>