adapter: Implement SHOW READYSET TABLES
Implement the SHOW READYSET TABLES command, using the response from the
new `table_statuses` RPC to construct a table with two columns: one
containing the name of the table, and one containing a string indicating
whether the table is snapshotting, snapshotted, or not replicated.

I've elected not to cover this with an integration test since it feels
overall low-value, and difficult enough to get tables in all three
statuses to test that it doesn't feel worth it to me.

Release-Note-Core: ReadySet now supports a custom `SHOW READYSET TABLES`
  SQL command, which will return a list of all tables ReadySet is aware
  of, along with their snapshotting status if they are being replicated
  by ReadySet, or "Not replicated" if ReadySet is not replicating them,
  either because the table is unsupported or because the table was not
  specified as part of `--replication-tables`.
Fixes: ENG-2077
Change-Id: I28034e538fcb1777479a6316db455a16f31994cb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4102
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>