Rename launchpad crate to readyset_util
This crate has been around for a *while*, and predates the renaming of a
bunch of our crates to have much more consistent and obvious naming. The
old "launchpad" name was cute, but didn't really explain what the point
of the crate was or what should go in there. Hopefully this should make
things a lot more obvious to people.

Change-Id: I926ae6e09835afa3c5430012f6a4e75b90cc14e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4104
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>