nom-sql: Parse <expr> <op> ANY|ALL|SOME (<expr>)
Add three new variants to Expr called OpAny, OpSome, and OpAll, and
parse all three. These fold a boolean operator between a left hand side
and all elements of an array on the right-hand side.

I went through a *lot* of iterations with parsing this as part of our
Pratt parser, including directly as a simple_expr (stack overflow,
unsurprisingly), as a special infix operator (couldn't get the
precedence right) and as a postfix operator (also couldn't get the
precedence right) - this approach, which matches what we do for
postgres-style `::` casts, is a little bit inelegant, but works
correctly and gives the correct precedence everywhere.

Currently, attempting to  actually lower these exprs to DfExpr returns
an unsupported error - actually evaluating this will come next.

Refs: ENG-2181
Change-Id: I2129217b4f5c8ca797f3a131f1c321f1df7a379c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4108
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>