dataflow-expression: Implement eval for expr op ANY/ALL (array)
Implement lowering and evaluation for the new `expr op ANY/ALL (array)`
expression, targeting a new pair of DfExpr variants.

Implementing this required factoring out the actual evaluation of binary
operators, since we need to do that in two places now - that's all just
moved around, nothing's actually changed in the impl there.

Also, coercion here is a little wonky, since we need to coerce the rhs
to an *array* of the lhs (if it isn't one already) - the more I think
about coercion in expr eval the more I'm thinking that nodes in
DfExpr *shouldn't* be responsible for coercing their inputs, and instead
we should emit coerce nodes at lowering time when we determine coercion
will be necessary.

Release-Note-Core: ReadySet now supports caching queries that use
  `<expr> <op> ANY|ALL|SOME (array)`
Fixes: ENG-2181
Change-Id: I3ee1098ef8351764e6042ffccfb3ec858b1fd01c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4114
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>