postgres: Raise replication error if NUMERIC scale > 28
Error out of replication if we attempt to deserialize a NUMERIC value
with a scale (i.e. # digits right of the decimal) greater than 28, as
this causes the panic seen in ENG-2187. Such values also get truncated
by our current implementation of NUMERIC via rust-decimal, and thus
become inconsistent with upstream.

Next step is to skip replication of tables for which the parsing of
replicator actions from the log encounters errors. Currently we only do
this when there is an error when applying said actions. As of now, this
change causes the replicator to fail and retry indefinitely, and the
offending table must be skipped manually.

Also, this currently only applies to scalars, and not elements in an
array, because of ENG-2254.

Refs: ENG-2230
Change-Id: Ib213c4bc90021616284de5486692e7e158c06019
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4124
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>